refactor: Move recipe handlers to separate file
